fix: integrate correctly with CODEOWNERS (#186)
CODEOWNERS does not support adding the `github-actions` bot into it,
but it supports adding other bots. But for a personal GitHub account,
bots cannot be added as collaborators. This effectively means, that
the only actual CODEOWNER with the power to approve PRs is the account
owner, in this case `akikanellis`. Therefore, this account is the only
one that can approve PRs, to achieve that, we use a PAT.

Fixes: #185